Extract transaction machine
